cram: review fixes for the C8 v3.0 writer
- doc.go: document that the CIGAR match operators =, X and M all
  encode as a base copy, so a =/X run re-reads as M — expected CRAM
  lossiness, not a rejected shape (the prior wording implied =/X were
  rejected).
- Add ITF-8/LTF-8 encoder round-trip tests covering the size-class
  boundaries and the int32/int64 extremes — the encoders were
  previously exercised only indirectly through the writer.
- FuzzRecordWriter now asserts per-record field equality after the
  round-trip, not just the record count.
- Drop a gratuitous int16 cast in the BAM-flags encode; comment the
  intentional twice-written block count in the slice header.

+++ pkg/htsgo/cram/writeencode.go
@@ -209,6 +209,8 @@ func encodeSliceHeader(f sliceHeaderFields) []byte {
 	b = appendITF8(b, f.alignmentSpan)
 	b = appendITF8(b, f.numRecords)
 	b = appendLTF8(b, f.recordCounter)
+	// The slice header carries the block count twice: the NumBlocks
+	// field and, separately, the length of the block-content-id array.
 	b = appendITF8(b, int32(len(f.contentIDs)))
 	b = appendITF8(b, int32(len(f.contentIDs)))
 	for _, id := range f.contentIDs {
@@ -355,7 +357,7 @@ func (e *recordEncoder) encodeRecord(rec *sam.Record) error {
 	// optimisation).
 	cf := int32(cfQualityPreserved | cfDetached)
 
-	b.bf = appendITF8(b.bf, int32(int16(rec.Flag)))
+	b.bf = appendITF8(b.bf, int32(rec.Flag))
 	b.cf = appendITF8(b.cf, cf)
 
 	refID := recordRefID(rec, e.refIndex)
